Keep external web API details out of tool wrappers
TinyFish-specific HTTP calls, response normalization, and API-key rotation now live behind an integration module so AI SDK tool files can stay focused on schemas, input validation, and adapter calls.

Constraint: Tool wrappers should expose the AI SDK surface without directly owning provider-specific network behavior.
Rejected: Keep TinyFish helpers under src/tools | that kept external API responsibilities mixed with tool registration code.
Confidence: high
Scope-risk: narrow
Reversibility: clean
Tested: pnpm run lint; pnpm run typecheck; pnpm run test
